+++ tests/Corvus.Text.Json.Tests.GeneratedModels.V4NonNullDefaulted/ObjectWithDefaultProperties.cs
@@ -1,0 +1,6 @@
+namespace Corvus.Text.Json.Tests.GeneratedModels.V4NonNullDefaulted.Draft202012;
+
+[Corvus.Json.JsonSchemaTypeGenerator("./object-with-default-properties-2020-12.json")]
+public readonly partial struct ObjectWithDefaultProperties
+{
+}